api/requirements/base.txt
Django==5.0.3
django-environ==0.11.2

djangorestframework==3.15.0
djangorestframework-jsonapi==6.1.0


django-extensions==3.2.3

psycopg2==2.9.9
psycopg2-binary==2.9.9


django-autocomplete-light==3.11.0

django-cors-headers==4.3.1


django-4-jet==1.0.9
six==1.16.0

django-filter==24.1


django-js-asset==2.2.0
django-mptt==0.16.0
django-oauth-toolkit==2.3.0

django-better-admin-arrayfield==1.4.2


django-querysetsequence==0.17
django-rest-framework-social-oauth2==1.2.0


